refactor: 컨테이너 설정하는 필드 및 로직을 추상클래스로 분리
