+++ src/test/java/com/api/cossc/ContainerInitialization.java
@@ -1,0 +1,23 @@
+package com.api.cossc;
+
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.MySQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+
+public abstract class ContainerInitialization {
+
+    @Container
+    public static MySQLContainer<?> mysql = new MySQLContainer<>("mysql:latest")
+            .withDatabaseName("cossc")
+            .withUsername("root")
+            .withPassword("root");
+
+    @DynamicPropertySource
+    public static void properties(DynamicPropertyRegistry registry) {
+        System.out.println("mysql.getJdbcUrl() = " + mysql.getJdbcUrl());
+        registry.add("spring.datasource.url", mysql::getJdbcUrl);
+        registry.add("spring.datasource.username", mysql::getUsername);
+        registry.add("spring.datasource.password", mysql::getPassword);
+    }
+}
